fixed typos and deleted kudos
the deleted text fits better "The program uses the Entrez module from [biopython](https://github.com/biopython/biopython). It was ported to python 3, and is slightly different from the original it was forked from. A big Kudos to the authors." fits better in home as special remark or under license